echoscript_backend_20250915_153911/echoscript_backend_20250915_153911/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
autoflake==2.3.1
av==13.1.0
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cryptography==45.0.5
ctranslate2==4.6.0
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
ecdsa==0.19.1
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.0
fastapi-crudrouter==0.8.6
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.18.0
flake8==7.3.0
flatbuffers==25.2.10
fonttools==4.59.0
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.33.4
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2
lightning==2.5.2
lightning-utilities==0.14.3
llvmlite==0.44.0
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
mypy==1.17.0
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
noisereduce==3.0.3
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.17.3
openai==1.97.0
optuna==4.4.0
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
primePy==1.3
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.9
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.2.3
pyreadline3==3.5.4
pytest==8.3.0
pytest-asyncio==0.21.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.20
pytorch-lightning==2.5.2
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
redis==5.0.8
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
semver==3.0.4
sendgrid==6.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
SQLAlchemy==2.0.36
starkbank-ecdsa==2.2.0
starlette==0.38.6
stripe==10.11.0
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.8.0+cpu
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0+cpu
torchmetrics==1.7.4
torchvision==0.23.0
tqdm==4.67.1
transformers==4.53.2
typer==0.16.0
types-cffi==1.17.0.20250523
types-fpdf2==2.8.3.20250516
types-passlib==1.7.7.20250602
types-pyasn1==0.6.0.20250516
types-pyOpenSSL==24.1.0.20240722
types-python-jose==3.5.0.20250531
types-redis==4.6.0.20241004
types-setuptools==80.9.0.20250529
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.6
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
